Implement functionality for card data submit

+++ src/main/java/rs/ac/ftn/uns/sep/bank/utils/dto/KpRequestDto.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 
 import javax.validation.constraints.Max;
+import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.time.LocalDateTime;
 
@@ -16,7 +17,7 @@ public class KpRequestDto {
     @Max(value = 100)
     private String merchantPassword;
 
-    private BigInteger amount;
+    private BigDecimal amount;
 
     private Integer merchantOrderId;
